robot-client-gui/requirements.txt
appdirs==1.4.3
autopep8==1.4.3
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Cython==0.29.6
enum34==1.1.6
Flask==1.0.2
gevent==1.4.0
greenlet==0.4.15
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
numpy==1.16.2
pep8==1.7.1
Pillow==5.4.1
pycodestyle==2.5.0
pygubu==0.9.8.2
requests==2.21.0
rope==0.14.0
scipy==1.2.1
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
